refactor: move HorizontalStyle and VerticalStyle composables to component package
